// callingMachine-master/src/app/interfaces/dashboard.interface.ts
export interface ICallingStatics {
  TodayNumberOfCalls: {
    departmentName: string;
    FaildCalledCount: string;
    SuccessfullCall: string;
  }[];
  YesterdayNumberOfCalls: {
    departmentName: string;
    FaildCalledCount: string;
    SuccessfullCall: string;
  }[];

  TotalCalls: string;
}